requirements.txt
pytorch-lightning==1.6.4
nltk==3.7
cmudict==1.0.2
wandb==0.12.19
pypinyin==0.46.0
pyter3==0.3
evaluate==0.2.2
jieba==0.42.1
Scrapy==2.7.1
scikit-learn==1.1.1
sacrebleu==2.1.0
matplotlib==3.5.2
pandas==1.4.2
sentencepiece==0.1.96
rouge-score==0.0.4
pynvml==11.4.1
seaborn==0.12.1